Improve caching, add logging, and fallback to whole document without args
Made it so caching stores files named uniquely per repo/branch so it can avoid accidentally posting the contents of another identically named file. Also made it log a bit nicer. Also made it so if you give it a file without line arguments it'll just post the entire file